fix(chat): fix sidebar infinity scroll spinner sticking (Issue #4752) (#4764)
